Tests/LibWeb/Text/expected/Editing/focus-reentrant-iframe-input-routing.txt

parent active: true
parent value: "X"
child value: ""
